added some mpi test
